Ensure we pass a ref to git fetch
Passing a rev doesn't always work since git can only fetch revs that
are pointed at by a ref (others may have been gc'd for example). But
we also validate in the task that the ref we fetched pointed at the
expected commit and otherwise abort; this means we bail out of tasks
early in the case that the branch was updated after the task was
scheduled but before it ran.