Fix for AllowPaymentRequest Bug
The solution is to snapshot the value of frame policy at the beginning of
a local_frame_client navigation and use it during document initialization.
If the frame owner is a remote one, the old value is used.

- capture frame policy at the start of navigation
    - local_frame_client_impl
    - NavigationInfo
    - render_frame_impl.BeginNavigationInternal

- pass it from renderer process to browser process through
  CommonNavigationParams in navigation_params.mojom
  - MakeCommonNavigationParams

- use value in CommonNavigationParams to init navigation_params
  - render_frame_impl.CommitNavigationInternal
  - web_navigation_params
    -> DocumentInit
    -> Document::InitializeFeaturePolicy

More detailed information could be found on crbug page.

A document_loader_test is added to further strengthen the test on frame
policy value used. Verified that the added test would fail without the
patch.

Bug: 972089
Change-Id: I694aa46b6e8e04bebc86b0e7886b856a9a302905