Add Web Platform Tests for the Split Cache
We add a web platform test for the cache when it is partitioned by top-
frame origin. The test caches a resource, then attempts to load it in
a popup window. With the split cache, if the popup window is
cross-origin, we would expect that request to be a cache miss. We
compare the number of requests that hit the server to our expectations.
To enable these tests, the existing framework is slightly refactored,
including to respond to CORS preflight requests.

The virtual test suites virtual/split-http-cache and
virtual/not-split-http-cache explicitly test the cases of a split and
not-split cache, respectively. Given that the split cache is not enabled
by default, the default test expectations are those of a not-split
cache.

Bug: 981970
Change-Id: I049fdaec4aebcbdeebdd3c77d13d6ca420d1bd62