Don't do the "simple display list" optimization when we have overflow clips.
The previous code tried to do it, but it did it wrongly, as the overflow clip
comes from the parent, not the child.

Thus when we change a style that influences it, we weren't invalidating the
SIMPLE_DISPLAY_LIST bit, and such.

Make the reftest that caught this fail more reliable.

Differential Revision: https://phabricator.services.mozilla.com/D51683

bugzilla-url: https://bugzilla.mozilla.org/show_bug.cgi?id=1590550
gecko-commit: e75d9dbb4853462d60ad18a589b67e379725c41f
gecko-integration-branch: autoland
gecko-reviewers: mattwoodrow