[AOM] Remove validation upon getting element refs
In order to comply with the current version of the spec, elements are
no longer validated on retrieval or getting. This means that explicitly
set elements are still able to be the attr-associated-element, even if
they have been moved into an invalid scope, such as a deeper shadow
root, or a different tree scope or removed entirely.

This is likely to change in future, but as the spec is still being
actively discussed it is better to comply with the current version.

BUG=981423

Change-Id: I496c06cfad5f056c49dedb3e9821781c755538b5